joined SCs fixtures
