--- raiden/tests/integration/contracts/conftest.py
@@ -1,3 +1,0 @@
-from raiden.tests.fixtures import *  # noqa
-from raiden.tests.integration.contracts.fixtures import *  # noqa
-from raiden.tests.integration.fixtures.blockchain import *  # noqa
