--- raiden/tests/integration/contracts/fixtures/__init__.py
@@ -1,1 +1,0 @@
-from .contracts import *  # noqa
